修改orderId类型为long

+++ src/main/java/com/zhuanche/serv/order/impl/DriverFeeDetailServiceImpl.java
@@ -44,7 +44,7 @@ public class DriverFeeDetailServiceImpl implements DriverFeeDetailService {
      * @param buyoutFlag 0-非一口价 1-一口价
      */
     @Override
-    public DriverCostDetailVO getDriverCostDetail(String orderNo, int orderId, Integer buyoutFlag) {
+    public DriverCostDetailVO getDriverCostDetail(String orderNo, long orderId, Integer buyoutFlag) {
         if (StringUtils.isBlank(orderNo) && orderId != 0) {
             return null;
         }
